+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.11 (2020-09-20)
 
 - added heist item data & stats
+- added heist item properties at the evaluate dialog
 - added heist item types at the market
 - added heist filter group at the market
 - updated the external link behaviour to skip in game notification